Move the instantiation code.
